requirements.txt
flask
requests
python-dotenv
sentence-transformers
scikit-learn
numpy
beautifulsoup4
